# backend/API/openapi.yml
openapi: 3.1.0
info:
  title: Training Management API
  description: API for managing training sessions, students, trainers, schedules, and plans
  version: v1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /trainers:
    get:
      tags:
      - people
      operationId: findAllTrainers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListTrainersDTO'
    post:
      tags:
      - people
      operationId: createTrainer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrainerDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ListTrainersDTO'
  /trainers/{trainerId}:
    get:
      tags:
      - people
      operationId: findTrainerById
      parameters:
      - name: trainerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ListTrainersDTO'
    put:
      tags:
      - people
      operationId: updateTrainer
      parameters:
      - name: trainerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrainerDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TrainerResponseDTO'
    delete:
      tags:
      - people
      operationId: deleteTrainer
      parameters:
      - name: trainerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  /schedules/{scheduleId}:
    get:
      tags:
      - scheduling
      operationId: getScheduleById
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TrainerScheduleResponseDTO'
    put:
      tags:
      - scheduling
      operationId: updateSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerScheduleRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TrainerScheduleResponseDTO'
    delete:
      tags:
      - scheduling
      operationId: deleteSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  /trainers/{trainerId}/schedules:
    get:
      tags:
      - scheduling
      operationId: getSchedulesByTrainer
      parameters:
      - name: trainerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainerScheduleResponseDTO'
    post:
      tags:
      - scheduling
      operationId: createSchedule
      parameters:
      - name: trainerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerScheduleRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TrainerScheduleResponseDTO'
  # ============================================
  # Scheduling Domain - Sessions
  # ============================================
...
  /students/{studentId}:
    get:
      tags:
      - people
      operationId: findStudentById
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentResponseDTO'
    put:
      tags:
      - people
      operationId: updateStudent
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentResponseDTO'
    delete:
      tags:
      - people
      operationId: deleteStudent
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /plans:
    get:
      tags:
      - plans
      operationId: getAllPlans
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentPlanResponseDTO'
    post:
      tags:
      - plans
      operationId: createPlan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentPlanRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentPlanResponseDTO'
  /plans/{planId}:
    get:
      tags:
      - plans
      operationId: getPlanById
      parameters:
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentPlanResponseDTO'
    delete:
      tags:
      - plans
      operationId: deletePlan
      parameters:
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  /plans/assignments/expiring-soon:
    get:
      tags:
      - analytics
      operationId: getExpiringSoonAssignments
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 7
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentPlanAssignmentResponseDTO'
...
  /trainers/{trainerId}/reset-password:
    patch:
      tags:
      - people
      operationId: resetPassword
      parameters:
      - name: trainerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetTrainerPasswordDTO'
        required: true
      responses:
        '200':
          description: OK
...
  /trainers/user-profile/{userId}:
    get:
      tags:
      - people
      operationId: findTrainerByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ListTrainersDTO'
  # ============================================
  # Analytics Domain
  # ============================================
...
  /students/{studentId}/plans/assignments:
    get:
      tags:
      - plans
      operationId: getStudentPlanHistory
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentPlanAssignmentResponseDTO'
  /students/{studentId}/plans/assignments/current:
    get:
      tags:
      - plans
      operationId: getCurrentStudentPlan
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                  $ref: '#/components/schemas/StudentPlanAssignmentResponseDTO'
...
components:
  schemas:
    CreateTrainerDTO:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        specialties:
          type: array
          items:
            type: string
        compensationType:
          type: string
          enum:
          - HOURLY
          - MONTHLY
    TrainerResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        specialties:
          type: array
          items:
            type: string
        compensationType:
          type: string
          enum:
          - HOURLY
          - MONTHLY
        joinDate:
          type: string
          format: date-time
        hoursWorked:
          type: integer
          format: int32
    TrainerScheduleRequestDTO:
      type: object
      properties:
        trainerId:
          type: string
          format: uuid
        weekday:
          type: integer
          format: int32
          maximum: 6
          minimum: 0
        startTime:
          type: string
        intervalDuration:
          type: integer
          format: int32
          minimum: 15
        seriesName:
          type: string
          minLength: 1
          pattern: ^(?!.*__).*$
      required:
      - seriesName
      - startTime
      - trainerId
      - weekday
    TrainerScheduleResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trainerId:
          type: string
          format: uuid
        weekday:
          type: integer
          format: int32
        startTime:
          type: string
        intervalDuration:
          type: integer
          format: int32
        seriesName:
          type: string
        effectiveFromTimestamp:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
        weekdayName:
          type: string
...
    AnamnesisRequestDTO:
      type: object
      properties:
        medication:
          type: string
        isDoctorAwareOfPhysicalActivity:
          type: boolean
        favoritePhysicalActivity:
          type: string
        hasInsomnia:
          type: string
          enum:
          - 'yes'
          - 'no'
          - sometimes
        dietOrientedBy:
          type: string
        cardiacProblems:
          type: string
        hasHypertension:
          type: boolean
        chronicDiseases:
          type: string
        difficultiesInPhysicalActivities:
          type: string
        medicalOrientationsToAvoidPhysicalActivity:
          type: string
        surgeriesInTheLast12Months:
          type: string
        respiratoryProblems:
          type: string
        jointMuscularBackPain:
          type: string
        spinalDiscProblems:
          type: string
        diabetes:
          type: string
        smokingDuration:
          type: string
        alteredCholesterol:
          type: boolean
        osteoporosisLocation:
          type: string
    PhysicalImpairmentRequestDTO:
      type: object
      properties:
        type:
          type: string
          enum:
          - visual
          - auditory
          - motor
          - intellectual
          - other
        name:
          type: string
          maxLength: 255
          minLength: 0
        observations:
          type: string
      required:
      - name
      - type
    StudentRequestDTO:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        name:
          type: string
          maxLength: 100
          minLength: 0
        surname:
          type: string
          maxLength: 100
          minLength: 0
        gender:
          type: string
          enum:
          - M
          - F
          - O
        birthDate:
          type: string
          format: date
        phone:
          type: string
          maxLength: 20
          minLength: 0
        profession:
          type: string
          maxLength: 100
          minLength: 0
        street:
          type: string
          maxLength: 255
          minLength: 0
        number:
          type: string
          maxLength: 10
          minLength: 0
        complement:
          type: string
          maxLength: 100
          minLength: 0
        neighborhood:
          type: string
          maxLength: 100
          minLength: 0
        cep:
          type: string
          maxLength: 10
          minLength: 0
        emergencyContactName:
          type: string
          maxLength: 100
          minLength: 0
        emergencyContactPhone:
          type: string
          maxLength: 20
          minLength: 0
        emergencyContactRelationship:
          type: string
          maxLength: 50
          minLength: 0
        objectives:
          type: string
        observations:
          type: string
        anamnesis:
          $ref: '#/components/schemas/AnamnesisRequestDTO'
        physicalImpairments:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalImpairmentRequestDTO'
      required:
      - birthDate
      - email
      - gender
      - name
      - surname
    AnamnesisResponseDTO:
      type: object
      properties:
        medication:
          type: string
        isDoctorAwareOfPhysicalActivity:
          type: boolean
        favoritePhysicalActivity:
          type: string
        hasInsomnia:
          type: string
          enum:
          - 'yes'
          - 'no'
          - sometimes
        dietOrientedBy:
          type: string
        cardiacProblems:
          type: string
        hasHypertension:
          type: boolean
        chronicDiseases:
          type: string
        difficultiesInPhysicalActivities:
          type: string
        medicalOrientationsToAvoidPhysicalActivity:
          type: string
        surgeriesInTheLast12Months:
          type: string
        respiratoryProblems:
          type: string
        jointMuscularBackPain:
          type: string
        spinalDiscProblems:
          type: string
        diabetes:
          type: string
        smokingDuration:
          type: string
        alteredCholesterol:
          type: boolean
        osteoporosisLocation:
          type: string
    PhysicalImpairmentResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - visual
          - auditory
          - motor
          - intellectual
          - other
        name:
          type: string
        observations:
          type: string
    StudentResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        surname:
          type: string
        gender:
          type: string
          enum:
          - M
          - F
          - O
        birthDate:
          type: string
          format: date
        phone:
          type: string
        profession:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: string
        neighborhood:
          type: string
        cep:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        emergencyContactRelationship:
          type: string
        objectives:
          type: string
        observations:
          type: string
        registrationDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        anamnesis:
          $ref: '#/components/schemas/AnamnesisResponseDTO'
        physicalImpairments:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalImpairmentResponseDTO'
...
    ListTrainersDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        specialties:
          type: array
          items:
            type: string
        compensationType:
          type: string
          enum:
          - HOURLY
          - MONTHLY
        active:
          type: boolean
        joinDate:
          type: string
          format: date-time
        hoursWorked:
          type: integer
          format: int32
...
    StudentPlanRequestDTO:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        maxDays:
          type: integer
          format: int32
        durationDays:
          type: integer
          format: int32
        description:
          type: string
      required:
      - durationDays
      - maxDays
      - name
    StudentPlanResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        maxDays:
          type: integer
          format: int32
        durationDays:
          type: integer
          format: int32
        description:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    StudentPlanAssignmentResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        studentName:
          type: string
        planId:
          type: string
          format: uuid
        planName:
          type: string
        planMaxDays:
          type: integer
          format: int32
        planDurationDays:
          type: integer
          format: int32
        durationDays:
          type: integer
          format: int32
        startDate:
          type: string
          format: date
        assignedByUserId:
          type: string
          format: uuid
        assignedByUserEmail:
          type: string
        assignmentNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        active:
          type: boolean
        expired:
          type: boolean
        expiringSoon:
          type: boolean
        daysRemaining:
          type: integer
          format: int64
...
    ResetTrainerPasswordDTO:
      type: object
      properties:
        newPassword:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
      - newPassword
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT