# backend/src/main/java/org/conexaotreinamento/conexaotreinamentobackend/controller/TrainerController.java
package org.conexaotreinamento.conexaotreinamentobackend.controller;

import java.util.List;
import java.util.UUID;

import org.conexaotreinamento.conexaotreinamentobackend.dto.request.CreateTrainerDTO;
import org.conexaotreinamento.conexaotreinamentobackend.dto.request.ResetTrainerPasswordDTO;
import org.conexaotreinamento.conexaotreinamentobackend.dto.response.ListTrainersDTO;
import org.conexaotreinamento.conexaotreinamentobackend.dto.response.TrainerResponseDTO;
import org.conexaotreinamento.conexaotreinamentobackend.service.TrainerService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/trainers")
@RequiredArgsConstructor
public class TrainerController {
    private final TrainerService trainerService;

    @PostMapping
    public ResponseEntity<ListTrainersDTO> createTrainerAndUser(@RequestBody @Valid CreateTrainerDTO request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(trainerService.create(request));
    }

    @GetMapping("/{trainerId}")
    public ResponseEntity<ListTrainersDTO> findTrainerById(@PathVariable UUID trainerId) {
        return ResponseEntity.ok(trainerService.findById(trainerId));
    }

    @GetMapping("/user-profile/{userId}")
    public ResponseEntity<ListTrainersDTO> findTrainerByUserId(@PathVariable UUID userId) {
        return ResponseEntity.ok(trainerService.findByUserId(userId));
    }

    @GetMapping
    public ResponseEntity<List<ListTrainersDTO>> findAllTrainers() {
        return ResponseEntity.ok(trainerService.findAll());
    }


    @PutMapping("/{trainerId}")
    public ResponseEntity<TrainerResponseDTO> updateTrainer(@PathVariable UUID trainerId, @RequestBody @Valid CreateTrainerDTO request) {
        return ResponseEntity.ok(trainerService.put(trainerId, request));
    }

    @DeleteMapping("/{trainerId}")
    public ResponseEntity<Void> deleteTrainer(@PathVariable UUID trainerId) {
        trainerService.delete(trainerId);
        return ResponseEntity.noContent().build();
    }

    @PatchMapping("/{trainerId}/reset-password")
    public ResponseEntity<Void> resetPassword(
            @PathVariable UUID trainerId,
            @RequestBody @Valid ResetTrainerPasswordDTO request
    ) {
        trainerService.resetPassword(trainerId, request.newPassword());
        return ResponseEntity.noContent().build();
    }
}